app/requirements.txt
# app
fastapi
uvicorn
pydantic
email-validator
uvicorn[standard]
jinja2
pytz
passlib[bcrypt]
python-jose[cryptography]
python-multipart

# db
SQLAlchemy
alembic
psycopg2-binary
asyncpg
